app/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
django-environ==0.4.5
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
docutils==0.17.1
drf-yasg==1.20.0
generics==3.3.1
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
# psycopg2==2.9.1
python-dotenv
psycopg2-binary
PyJWT==2.1.0
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
# ruamel.yaml==0.17.10
# ruamel.yaml.clib==0.2.6
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
cloudinary==1.26.0
